Added the script "GenerateRandomTrialPoints.m" to create a set of 100 random trial points.
This script works as follows (Taken from the explanation given in the script header)

This script generates a random set of trial points from which we can pick
one to start the algorithm. It's useful in two ways. One is that we don't
have to spend the computational effort in generating random trial points
every time we run the algorithm. The other is that we can for instance
start with the same trial point every time, and run controlled
experiments.

Logic:
We need to remove a certain number of vehicles from the network. We do this by
randomly choosing a few OD pairs, and then remove 25 % of the demand from
each of the OD pair.

In the end this script just saves the third column of the OD pair table.
It assumes that the order of OD pairs is maintained in the rest of the
script.