Added a queuing model with Fsolve
Extracted data and aimsun file from the multi model SO code, and then
got a working version of the QFsolve code for the small network. It
doesn't work for the larger network, yet and gives NAN values. Could be
an issue with the code, or with the data I extracted.